# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: The Diner - Restaurant Operations Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/recent-orders:
    get:
      operationId: getDashboardRecentOrders
      tags: [dashboard]
      summary: Get recent orders for dashboard
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /dashboard/top-items:
    get:
      operationId: getDashboardTopItems
      tags: [dashboard]
      summary: Get top selling menu items
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 4
      responses:
        "200":
          description: Top menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopItem"
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of customers with stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerWithStats"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer with order history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail with order history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ─── SETTINGS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    TopItem:
      type: object
      required: [id, name, orderCount, categoryName, price, available]
      properties:
        id:
          type: integer
        name:
          type: string
        orderCount:
          type: integer
        categoryName:
          type: string
        price:
          type: number
        available:
          type: boolean
        imageKey:
          type: ["string", "null"]
...
    OrderItem:
      type: object
      required: [menuItemId, name, quantity, unitPrice]
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
    Order:
      type: object
      required: [id, status, total, createdAt, items]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, preparing, ready, fulfilled, cancelled]
        total:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...
    Customer:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CustomerWithStats:
      type: object
      required: [id, name, orderCount, totalSpend, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        orderCount:
          type: integer
        totalSpend:
          type: number
        lastOrderDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    CustomerDetail:
      type: object
      required: [id, name, orderCount, totalSpend, createdAt, orders]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        orderCount:
          type: integer
        totalSpend:
          type: number
        lastOrderDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]

    # ─── SETTINGS SCHEMAS ────────────────────────────────────────
...