# packages/types/src/generated/types/customerWithStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Diner - Restaurant Operations Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerWithStats {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  orderCount: number;
  totalSpend: number;
  /** @nullable */
  lastOrderDate?: string | null;
  createdAt: Date;
}