# packages/types/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Diner - Restaurant Operations Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt?: Date;
  items: OrderItem[];
}