# packages/types/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Diner - Restaurant Operations Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  menuItemId: number;
  name: string;
  quantity: number;
  unitPrice: number;
}